+++ apps/core/admin.py
@@ -2,6 +2,7 @@ from django.contrib import admin
 
 from apps.core.models import (
     EmailSent,
+    PageCrawlWork,
     Project,
     ProjectStateTransition,
     ProjectSyncRequest,
@@ -15,3 +16,4 @@ admin.site.register(ProjectStateTransition)
 admin.site.register(ProjectSyncRequest)
 admin.site.register(SitemapInventory)
 admin.site.register(SitemapCandidate)
+admin.site.register(PageCrawlWork)